microservices_backend/requirements.txt
langchain
langchain-core
langchain-groq
langgraph
python-dotenv
fastapi
uvicorn
langchain-huggingface
sentence_transformers